Add files via upload
data.json and data1.json are the files which contains words and their meaning. One can create their own or can download from web.
 Version 1.1 Code.html and  Version 1.2 Code.html are supportive files which helps to import data from json file and which helps to run our project online also.
Volcanoes.txt is a simple txt file which contains a data which provides css type for dictionary.
app1.py is the main file which is created in python. Run this file to execute our project. Simply add all these files to workspace of editor and run the dictionary, excluding any of this file we cannot run the project
This project is ideal to understand python in advance.